+++ app/controllers/mokio/application_controller.rb
@@ -1,3 +1,6 @@
 class Mokio::ApplicationController < ActionController::Base
   include Mokio::Concerns::Controllers::Application
+  include Mokio::HeadersHelper
+
+  before_action :set_no_cache_headers
 end