+++ db/migrate/20200216211300_add_session_token_to_mokio_users.rb
@@ -1,0 +1,5 @@
+class AddSessionTokenToMokioUsers < ActiveRecord::Migration[6.0]
+  def change
+    add_column :mokio_users, :session_token, :string
+  end
+end